Equipment + Inventory
Trying to be efficient.. but just not happening. I need to be able to "construct" a state prop to render it as a value.

character.equipment.[item]

where item => ["head", "back"]

It works for the rest of the component, just not the callback to render.

Added Match() function to hold me over and get me off the suicide watch list on this issue.