Creator Talents + Money
Minor UI updates around spacing.

So.. talents are done. 1AM coding will do that. Not sure what to do with the ATK and/or DMG "bonuses" you get from the talents.. that's a little hard to build logic around as it's multi-faceted.

Still thinking on items, their content models, and how they contribute to the character sheet. Hacky HP solution is place.. works well but not sure I like it.