Creator
Temporary buff expanded into attributes. With this Attributes now show total instead of just base stat. This impacts Attack, Armor, and Modifier.

Utility functions expanded to account for all the above.

Going to chalk this up as a major win, been thinking on this for a while.

Minor updates to Header, Footer, and Input.