Augments + Temporary
With the expansion of Temporary and the addition of Augments via functions all over Creator we need to update + add to the character state/object when loading a character without the new objects.

Once the character has the new objects within everything falls back to a basic copy object onto state. Still not 100% ideal, thinking about a better solution.

TLDR: Essentially the new functions are calling key/value pairs that don't exist in older character objects. The solve is to either update the object with everything that could be needed on loading a character -OR- update every "new" reference to a ternary. The ladder causes other issues as I use map() to output many of the configurable stats on a character.