Temporary + Augments
Final update, make sure when loading an existing character that we check for the new objects and if not there we take the existing data and merge it with the new objects that are assigned via CharacterUpdate state.

As before, if the character has all the new properties it's a simple overlaying of the object on to the state.